Begin some work to create a suite of containerized tools.
Move update-pipenv-lock to a dev/ directory, to start a trend of keeping scripts there.

Modify the Dockerfile such that it can have a dev instance and a production instance.

Remove cargo and rust, as ruff no longer needs them (This saves about 600 MB off of the size of the container)